+++ aragog_cli/src/migration.rs
@@ -29,7 +29,10 @@ impl Migration {
     pub fn migration_path(schema_path: &str) -> Result<String, MigrationError> {
         let db_path = format!("{}/{}", schema_path, MIGRATION_PATH);
         if !Path::new(&db_path).is_dir() {
-            println!("{} Missing {}/ path. creating it..", MIGRATION_PATH, LOG_STR);
+            println!(
+                "{} Missing {}/ path. creating it..",
+                MIGRATION_PATH, LOG_STR
+            );
             fs::create_dir(&db_path)?;
         }
         Ok(db_path)
@@ -40,7 +43,12 @@ impl Migration {
         let data_str = serde_yaml::to_string(&data).unwrap();
         let version = Utc::now().timestamp_millis() as u64;
         let migration_path = Self::migration_path(schema_path)?;
-        let path = format!("{}/{}_{}.yaml", migration_path, version, name.to_ascii_lowercase());
+        let path = format!(
+            "{}/{}_{}.yaml",
+            migration_path,
+            version,
+            name.to_ascii_lowercase()
+        );
         let mut file = File::create(&path)?;
         let buff = format!("{}{}", HELP_MESSAGE, data_str);
         file.write_all(buff.as_bytes())?;