+++ src/schema/mod.rs
@@ -11,7 +11,8 @@ mod database_schema;
 mod graph_schema;
 mod index_schema;
 
-pub(crate) const SCHEMA_DEFAULT_PATH: &str = "./src/config/db/schema.yaml";
+/// Default schema path, can be overridden manually or set as `SCHEMA_PATH` env var
+pub const SCHEMA_DEFAULT_PATH: &str = "./src/config/db/schema.yaml";
 
 /// Trait used for all schema elements allowing to synchronise schema changes.
 /// Used by `aragog_cli` for migrations and `DatabaseConnectionPool`